Make compile_func_to_ir() a Function object method
Upcoming patches will convert the function bodies to a proper
`Expression` field, and this change moves things closer to where
they'll be.
