+++ edb/edgeql/compiler/__init__.py
@@ -131,7 +131,6 @@ from typing import *
 from edb import errors
 
 from edb.edgeql import parser as ql_parser
-from edb.edgeql import qltypes
 
 from edb.common import debug
 from edb.common import markup  # NOQA
@@ -453,76 +452,6 @@ def evaluate_ast_to_python_val(
     """
     ir = compile_ast_fragment_to_ir(tree, schema, modaliases=modaliases)
     return ireval.evaluate_to_python_val(ir.expr, schema=ir.schema)
-
-
-def compile_func_to_ir(
-    func: s_func.Function,
-    schema: s_schema.Schema,
-) -> irast.Statement:
-    """Compile an EdgeQL function into EdgeDB IR.
-
-    Args:
-        func:
-            A function object.
-
-        schema:
-            A schema instance where the function is defined.
-
-    Returns:
-        An instance of :class:`ir.ast.Statement` representing the
-        function body.
-    """
-    if debug.flags.edgeql_compile:
-        debug.header('EdgeQL Function')
-        debug.print(func.get_code(schema))
-
-    code = func.get_code(schema)
-    assert code is not None
-    trees = ql_parser.parse_block(code + ';')
-    if len(trees) != 1:
-        raise errors.InvalidFunctionDefinitionError(
-            'functions can only contain one statement')
-
-    tree = trees[0]
-
-    param_anchors = s_func.get_params_symtable(
-        func.get_params(schema), schema,
-        inlined_defaults=func.has_inlined_defaults(schema))
-
-    ir = compile_ast_to_ir(
-        tree, schema,
-        anchors=param_anchors,
-        func_params=func.get_params(schema),
-        # the body of a session_only function can contain calls to
-        # other session_only functions
-        session_mode=func.get_session_only(schema),
-    )
-
-    assert isinstance(ir, irast.Statement)
-    schema = ir.schema
-
-    return_type = func.get_return_type(schema)
-    if (not ir.stype.issubclass(schema, return_type)
-            and not ir.stype.implicitly_castable_to(return_type, schema)):
-        raise errors.InvalidFunctionDefinitionError(
-            f'return type mismatch in function declared to return '
-            f'{return_type.get_verbosename(schema)}',
-            details=f'Actual return type is '
-                    f'{ir.stype.get_verbosename(schema)}',
-            context=tree.context,
-        )
-
-    return_typemod = func.get_return_typemod(schema)
-    if (return_typemod is not qltypes.TypeModifier.SET_OF
-            and ir.cardinality.is_multi()):
-        raise errors.InvalidFunctionDefinitionError(
-            f'return cardinality mismatch in function declared to return '
-            f'a singleton',
-            details=f'Function may return a set with more than one element.',
-            context=tree.context,
-        )
-
-    return ir
 
 
 def compile_constant_tree_to_ir(
